Moves preComputeArtDirectives to compute as a private method (it simplfies implmentation in various frameworks).
